// src/lib/api/models/CreateInfraDependencyRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UUID } from './UUID';
export type CreateInfraDependencyRequest = {
    config?: Record<string, any>;
    dependencyName?: string;
    dependencyType?: string;
    hostServerTypeId?: UUID;
    platformTypeId?: UUID;
};